Move about panel display to a class-based approach
Instead of manually setting display characteristics in JS, we now have
a class in CSS for "closed".  The button simply toggles the presence of
this class.  Furthermore, we can use CSS3 animations to remove the need
for JS-based animations, and let the browser handle the tricky bits of
canceling animations, etc.