Refactor starmap to be reactive
We now use a virtually-sized SVG canvas of [columns*100, rows*100], set
the viewBox on the SVG element to the actual size of the grid that fits
in that canvas (automatically determining the cell size), and let the
browser handle the scaling.

Additionally, the items inside each star are positioned relative to its
group, and not globally, and then the group itself it translated around.